+++ src/Editors/ReplaceRequest.php
@@ -10,6 +10,6 @@ final readonly class ReplaceRequest
     public function __construct(
         public string $find,
         public string $replace,
-        public int $options = SearchOptions::CaseSensitive | SearchOptions::ReplaceAll,
+        public int $options = SearchOptions::CaseSensitive | SearchOptions::DoReplace | SearchOptions::ReplaceAll,
     ) {}
 }